pyproject.toml
[project]
name = "weird-gen"
version = "0.1.0"
description = "Replication of Weird Generalization & Inductive Backdoors (Betley et al., 2025)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core ML
    "torch>=2.2.0",
    "transformers>=4.45.0",
    "peft>=0.13.0",
    "datasets>=3.0.0",  # Also used for fetching Alpaca/GSM8K
    "accelerate>=1.0.0",
    "bitsandbytes>=0.44.0",
    "trl>=0.12.0",  # For SFTTrainer
    "unsloth",  # 2x faster LoRA training
    
    # Inference
    "vllm>=0.6.0",
    
    # API clients (for data generation & judging)
    "openai>=1.50.0",
    "anthropic>=0.37.0",
    
    # Data processing
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    
    # Visualization
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
    
    # Config & CLI
    "pyyaml>=6.0.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    
    # Experiment tracking
    "wandb>=0.18.0",
    
    # Utilities
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.0",
    
    # HTTP client (for OpenRouter)
    "httpx>=0.27.0",
    
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.6.0",
    "ipython>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]  # Line length handled separately

[tool.hatch.build.targets.wheel]
packages = ["scripts"]
